docs(runtime): re-attach the finally lane doc comments to their items
the extraction left two orphaned doc comments in task.rs (the const and
the preview helper's) · they ride with their items in task/finally.rs.
